+++ packages/rum/src/boot/startRecording.spec.ts
@@ -46,8 +46,6 @@ describe('startRecording', () => {
             },
             view: {
               id: viewId,
-              referrer: '',
-              url: 'http://example.org',
             },
           }
         },